refactor(domain): Event 및 Campaign 도메인 모델 개선
실제 비즈니스 요구사항을 반영하여 도메인 모델을 개선하고, 관련된 코드 구조를 리팩토링합니다.

- `Event` 엔티티에 `Campaign`과의 연관관계를 설정하고, `id` 타입을 `Long`으로 변경
- `Kafka_ProducerDto`의 `eventId` 타입을 `Long`으로 변경하여 도메인 모델과 일치시킴
- `EventRepository`를 `domain`에서 `repository` 패키지로 이동
- 향후 참여(Participation) 도메인 로직을 위한 패키지 구조 생성

+++ common-messaging/src/main/java/com/axon/messaging/dto/Kafka_ProducerDto.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class Kafka_ProducerDto {
     private CampaignType campaignType;
-    private int eventId;
+    private Long eventId;
     private int userId;
     private int productId;
     private Long timestamp;
